+++ internal/purchase/mocks_test.go
@@ -2,17 +2,20 @@ package purchase
 
 import (
 	"context"
-	"time"
 
-	"github.com/LeanerCloud/CUDly/internal/config"
 	"github.com/LeanerCloud/CUDly/internal/credentials"
 	"github.com/LeanerCloud/CUDly/internal/email"
+	"github.com/LeanerCloud/CUDly/internal/mocks"
 	"github.com/LeanerCloud/CUDly/pkg/common"
 	"github.com/LeanerCloud/CUDly/pkg/provider"
 	"github.com/aws/aws-sdk-go-v2/service/sts"
-	"github.com/jackc/pgx/v5"
 	"github.com/stretchr/testify/mock"
 )
+
+// MockConfigStore is the shared testify mock for config.StoreInterface.
+// All Fn-override fields (GetPlanAccountsFn, SavePurchaseExecutionFn, etc.)
+// and default behaviours live in internal/mocks.
+type MockConfigStore = mocks.MockConfigStore
 
 // MockProviderFactory is a mock implementation of ProviderFactoryInterface
 type MockProviderFactory struct {
@@ -153,477 +156,6 @@ func (m *MockServiceClient) GetValidResourceTypes(ctx context.Context) ([]string
 	}
 	return args.Get(0).([]string), args.Error(1)
 }
-
-// MockConfigStore is a mock implementation of config.StoreInterface
-type MockConfigStore struct {
-	mock.Mock
-	// GetPlanAccountsFn overrides GetPlanAccounts when non-nil (used in multi-account tests).
-	GetPlanAccountsFn func(ctx context.Context, planID string) ([]config.CloudAccount, error)
-	// SavePurchaseExecutionFn overrides SavePurchaseExecution when non-nil.
-	SavePurchaseExecutionFn func(ctx context.Context, exec *config.PurchaseExecution) error
-}
-
-func (m *MockConfigStore) GetGlobalConfig(ctx context.Context) (*config.GlobalConfig, error) {
-	args := m.Called(ctx)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).(*config.GlobalConfig), args.Error(1)
-}
-
-func (m *MockConfigStore) SaveGlobalConfig(ctx context.Context, cfg *config.GlobalConfig) error {
-	args := m.Called(ctx, cfg)
-	return args.Error(0)
-}
-
-func (m *MockConfigStore) GetServiceConfig(ctx context.Context, provider, service string) (*config.ServiceConfig, error) {
-	args := m.Called(ctx, provider, service)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).(*config.ServiceConfig), args.Error(1)
-}
-
-func (m *MockConfigStore) SaveServiceConfig(ctx context.Context, cfg *config.ServiceConfig) error {
-	args := m.Called(ctx, cfg)
-	return args.Error(0)
-}
-
-func (m *MockConfigStore) ListServiceConfigs(ctx context.Context) ([]config.ServiceConfig, error) {
-	args := m.Called(ctx)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).([]config.ServiceConfig), args.Error(1)
-}
-
-func (m *MockConfigStore) CreatePurchasePlan(ctx context.Context, plan *config.PurchasePlan) error {
-	args := m.Called(ctx, plan)
-	return args.Error(0)
-}
-
-func (m *MockConfigStore) GetPurchasePlan(ctx context.Context, planID string) (*config.PurchasePlan, error) {
-	args := m.Called(ctx, planID)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).(*config.PurchasePlan), args.Error(1)
-}
-
-func (m *MockConfigStore) UpdatePurchasePlan(ctx context.Context, plan *config.PurchasePlan) error {
-	args := m.Called(ctx, plan)
-	return args.Error(0)
-}
-
-func (m *MockConfigStore) IncrementPlanCurrentStep(ctx context.Context, planID string) error {
-	args := m.Called(ctx, planID)
-	return args.Error(0)
-}
-
-// UpdatePurchasePlanTx falls back to UpdatePurchasePlan when no
-// expectation is registered. Mirrors the pattern used by
-// SavePurchaseExecutionTx and the api-package mock.
-func (m *MockConfigStore) UpdatePurchasePlanTx(ctx context.Context, tx pgx.Tx, plan *config.PurchasePlan) error {
-	for _, call := range m.ExpectedCalls {
-		if call.Method == "UpdatePurchasePlanTx" {
-			args := m.Called(ctx, tx, plan)
-			return args.Error(0)
-		}
-	}
-	return m.UpdatePurchasePlan(ctx, plan)
-}
-
-func (m *MockConfigStore) DeletePurchasePlan(ctx context.Context, planID string) error {
-	args := m.Called(ctx, planID)
-	return args.Error(0)
-}
-
-func (m *MockConfigStore) ListPurchasePlans(ctx context.Context, filter config.PurchasePlanFilter) ([]config.PurchasePlan, error) {
-	args := m.Called(ctx, filter)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).([]config.PurchasePlan), args.Error(1)
-}
-
-func (m *MockConfigStore) SavePurchaseExecution(ctx context.Context, exec *config.PurchaseExecution) error {
-	if m.SavePurchaseExecutionFn != nil {
-		return m.SavePurchaseExecutionFn(ctx, exec)
-	}
-	args := m.Called(ctx, exec)
-	return args.Error(0)
-}
-
-func (m *MockConfigStore) GetPendingExecutions(ctx context.Context) ([]config.PurchaseExecution, error) {
-	args := m.Called(ctx)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).([]config.PurchaseExecution), args.Error(1)
-}
-
-func (m *MockConfigStore) GetExecutionsByStatuses(ctx context.Context, statuses []string, limit int) ([]config.PurchaseExecution, error) {
-	args := m.Called(ctx, statuses, limit)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).([]config.PurchaseExecution), args.Error(1)
-}
-
-func (m *MockConfigStore) GetPlannedExecutions(ctx context.Context, statuses []string, limit int) ([]config.PurchaseExecution, error) {
-	args := m.Called(ctx, statuses, limit)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).([]config.PurchaseExecution), args.Error(1)
-}
-
-func (m *MockConfigStore) GetStaleApprovedExecutions(ctx context.Context, olderThan time.Duration) ([]config.PurchaseExecution, error) {
-	args := m.Called(ctx, olderThan)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).([]config.PurchaseExecution), args.Error(1)
-}
-
-func (m *MockConfigStore) GetExecutionByID(ctx context.Context, executionID string) (*config.PurchaseExecution, error) {
-	args := m.Called(ctx, executionID)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).(*config.PurchaseExecution), args.Error(1)
-}
-
-func (m *MockConfigStore) GetExecutionByPlanAndDate(ctx context.Context, planID string, scheduledDate time.Time) (*config.PurchaseExecution, error) {
-	args := m.Called(ctx, planID, scheduledDate)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).(*config.PurchaseExecution), args.Error(1)
-}
-
-func (m *MockConfigStore) CleanupOldExecutions(ctx context.Context, retentionDays int) (int64, error) {
-	args := m.Called(ctx, retentionDays)
-	return args.Get(0).(int64), args.Error(1)
-}
-
-func (m *MockConfigStore) CountPendingExecutionsForAccount(ctx context.Context, accountID string) (int, error) {
-	args := m.Called(ctx, accountID)
-	return args.Int(0), args.Error(1)
-}
-
-func (m *MockConfigStore) ListPendingExecutionIDsForAccount(ctx context.Context, accountID string) ([]string, error) {
-	args := m.Called(ctx, accountID)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).([]string), args.Error(1)
-}
-
-func (m *MockConfigStore) TransitionExecutionStatus(ctx context.Context, executionID string, fromStatuses []string, toStatus string) (*config.PurchaseExecution, error) {
-	args := m.Called(ctx, executionID, fromStatuses, toStatus)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).(*config.PurchaseExecution), args.Error(1)
-}
-
-func (m *MockConfigStore) ListStuckExecutions(ctx context.Context, statuses []string, olderThan time.Duration) ([]config.PurchaseExecution, error) {
-	args := m.Called(ctx, statuses, olderThan)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).([]config.PurchaseExecution), args.Error(1)
-}
-
-func (m *MockConfigStore) GetScheduledExecutionsDue(ctx context.Context) ([]config.PurchaseExecution, error) {
-	args := m.Called(ctx)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).([]config.PurchaseExecution), args.Error(1)
-}
-
-func (m *MockConfigStore) SavePurchaseHistory(ctx context.Context, record *config.PurchaseHistoryRecord) error {
-	args := m.Called(ctx, record)
-	return args.Error(0)
-}
-
-func (m *MockConfigStore) GetPurchaseHistory(ctx context.Context, accountID string, limit int) ([]config.PurchaseHistoryRecord, error) {
-	args := m.Called(ctx, accountID, limit)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).([]config.PurchaseHistoryRecord), args.Error(1)
-}
-
-func (m *MockConfigStore) GetAllPurchaseHistory(ctx context.Context, limit int) ([]config.PurchaseHistoryRecord, error) {
-	args := m.Called(ctx, limit)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).([]config.PurchaseHistoryRecord), args.Error(1)
-}
-
-func (m *MockConfigStore) GetActivePurchaseHistory(ctx context.Context, asOf time.Time) ([]config.PurchaseHistoryRecord, error) {
-	args := m.Called(ctx, asOf)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).([]config.PurchaseHistoryRecord), args.Error(1)
-}
-
-func (m *MockConfigStore) GetPurchaseHistoryFiltered(ctx context.Context, filter config.PurchaseHistoryFilter) ([]config.PurchaseHistoryRecord, error) {
-	args := m.Called(ctx, filter)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).([]config.PurchaseHistoryRecord), args.Error(1)
-}
-
-func (m *MockConfigStore) GetPurchaseHistoryByPurchaseID(ctx context.Context, purchaseID string) (*config.PurchaseHistoryRecord, error) {
-	args := m.Called(ctx, purchaseID)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).(*config.PurchaseHistoryRecord), args.Error(1)
-}
-
-func (m *MockConfigStore) MarkPurchaseRevoked(ctx context.Context, purchaseID string, revokedAt time.Time, revokedVia string, supportCaseID string) error {
-	args := m.Called(ctx, purchaseID, revokedAt, revokedVia, supportCaseID)
-	return args.Error(0)
-}
-
-func (m *MockConfigStore) SaveRIExchangeRecord(ctx context.Context, record *config.RIExchangeRecord) error {
-	args := m.Called(ctx, record)
-	return args.Error(0)
-}
-
-func (m *MockConfigStore) GetRIExchangeRecord(ctx context.Context, id string) (*config.RIExchangeRecord, error) {
-	args := m.Called(ctx, id)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).(*config.RIExchangeRecord), args.Error(1)
-}
-
-func (m *MockConfigStore) GetRIExchangeRecordByToken(ctx context.Context, token string) (*config.RIExchangeRecord, error) {
-	args := m.Called(ctx, token)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).(*config.RIExchangeRecord), args.Error(1)
-}
-
-func (m *MockConfigStore) GetRIExchangeHistory(ctx context.Context, since time.Time, limit int) ([]config.RIExchangeRecord, error) {
-	args := m.Called(ctx, since, limit)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).([]config.RIExchangeRecord), args.Error(1)
-}
-
-func (m *MockConfigStore) TransitionRIExchangeStatus(ctx context.Context, id string, fromStatus string, toStatus string) (*config.RIExchangeRecord, error) {
-	args := m.Called(ctx, id, fromStatus, toStatus)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).(*config.RIExchangeRecord), args.Error(1)
-}
-
-func (m *MockConfigStore) CompleteRIExchange(ctx context.Context, id string, exchangeID string) error {
-	args := m.Called(ctx, id, exchangeID)
-	return args.Error(0)
-}
-
-func (m *MockConfigStore) FailRIExchange(ctx context.Context, id string, errorMsg string) error {
-	args := m.Called(ctx, id, errorMsg)
-	return args.Error(0)
-}
-
-func (m *MockConfigStore) StampRIExchangeApprovedBy(ctx context.Context, id string, approverEmail string) error {
-	args := m.Called(ctx, id, approverEmail)
-	return args.Error(0)
-}
-
-func (m *MockConfigStore) GetRIExchangeDailySpend(ctx context.Context, date time.Time) (string, error) {
-	args := m.Called(ctx, date)
-	return args.String(0), args.Error(1)
-}
-
-func (m *MockConfigStore) CancelAllPendingExchanges(ctx context.Context) (int64, error) {
-	args := m.Called(ctx)
-	return args.Get(0).(int64), args.Error(1)
-}
-
-func (m *MockConfigStore) GetStaleProcessingExchanges(ctx context.Context, olderThan time.Duration) ([]config.RIExchangeRecord, error) {
-	args := m.Called(ctx, olderThan)
-	if args.Get(0) == nil {
-		return nil, args.Error(1)
-	}
-	return args.Get(0).([]config.RIExchangeRecord), args.Error(1)
-}
-
-func (m *MockConfigStore) CreateCloudAccount(ctx context.Context, account *config.CloudAccount) error {
-	return nil
-}
-func (m *MockConfigStore) GetCloudAccount(ctx context.Context, id string) (*config.CloudAccount, error) {
-	// Route through m.Called only when a test has registered an
-	// expectation; otherwise fall back to the historical "no account"
-	// stub so the dozens of pre-existing tests that don't care about
-	// account lookups stay green. Mirrors the isExpected pattern in
-	// internal/api/mocks_test.go.
-	for _, call := range m.ExpectedCalls {
-		if call.Method == "GetCloudAccount" {
-			args := m.Called(ctx, id)
-			if args.Get(0) == nil {
-				return nil, args.Error(1)
-			}
-			return args.Get(0).(*config.CloudAccount), args.Error(1)
-		}
-	}
-	return nil, nil
-}
-func (m *MockConfigStore) GetCloudAccountByExternalID(ctx context.Context, provider, externalID string) (*config.CloudAccount, error) {
-	for _, call := range m.ExpectedCalls {
-		if call.Method == "GetCloudAccountByExternalID" {
-			args := m.Called(ctx, provider, externalID)
-			if args.Get(0) == nil {
-				return nil, args.Error(1)
-			}
-			return args.Get(0).(*config.CloudAccount), args.Error(1)
-		}
-	}
-	return nil, nil
-}
-func (m *MockConfigStore) UpdateCloudAccount(ctx context.Context, account *config.CloudAccount) error {
-	return nil
-}
-func (m *MockConfigStore) DeleteCloudAccount(ctx context.Context, id string) error {
-	return nil
-}
-func (m *MockConfigStore) ListCloudAccounts(ctx context.Context, filter config.CloudAccountFilter) ([]config.CloudAccount, error) {
-	return nil, nil
-}
-func (m *MockConfigStore) SaveAccountCredential(ctx context.Context, accountID, credentialType, encryptedBlob string) error {
-	return nil
-}
-func (m *MockConfigStore) GetAccountCredential(ctx context.Context, accountID, credentialType string) (string, error) {
-	return "", nil
-}
-func (m *MockConfigStore) DeleteAccountCredentials(ctx context.Context, accountID string) error {
-	return nil
-}
-func (m *MockConfigStore) HasAccountCredentials(ctx context.Context, accountID string) (bool, error) {
-	return false, nil
-}
-func (m *MockConfigStore) GetAccountServiceOverride(ctx context.Context, accountID, provider, service string) (*config.AccountServiceOverride, error) {
-	return nil, nil
-}
-func (m *MockConfigStore) SaveAccountServiceOverride(ctx context.Context, override *config.AccountServiceOverride) error {
-	return nil
-}
-func (m *MockConfigStore) DeleteAccountServiceOverride(ctx context.Context, accountID, provider, service string) error {
-	return nil
-}
-func (m *MockConfigStore) ListAccountServiceOverrides(ctx context.Context, accountID string) ([]config.AccountServiceOverride, error) {
-	return nil, nil
-}
-func (m *MockConfigStore) SetPlanAccounts(ctx context.Context, planID string, accountIDs []string) error {
-	return nil
-}
-func (m *MockConfigStore) GetPlanAccounts(ctx context.Context, planID string) ([]config.CloudAccount, error) {
-	if m.GetPlanAccountsFn != nil {
-		return m.GetPlanAccountsFn(ctx, planID)
-	}
-	return nil, nil
-}
-
-func (m *MockConfigStore) CreateAccountRegistration(_ context.Context, _ *config.AccountRegistration) error {
-	return nil
-}
-func (m *MockConfigStore) GetAccountRegistration(_ context.Context, _ string) (*config.AccountRegistration, error) {
-	return nil, nil
-}
-func (m *MockConfigStore) GetAccountRegistrationByToken(_ context.Context, _ string) (*config.AccountRegistration, error) {
-	return nil, nil
-}
-func (m *MockConfigStore) ListAccountRegistrations(_ context.Context, _ config.AccountRegistrationFilter) ([]config.AccountRegistration, error) {
-	return nil, nil
-}
-func (m *MockConfigStore) UpdateAccountRegistration(_ context.Context, _ *config.AccountRegistration) error {
-	return nil
-}
-func (m *MockConfigStore) TransitionRegistrationStatus(_ context.Context, _ *config.AccountRegistration, _ string) error {
-	return nil
-}
-func (m *MockConfigStore) DeleteAccountRegistration(_ context.Context, _ string) error {
-	return nil
-}
-
-// ── Purchase suppressions (Commit 2 of bulk-purchase-with-grace)
-func (m *MockConfigStore) CreateSuppression(_ context.Context, _ *config.PurchaseSuppression) error {
-	return nil
-}
-func (m *MockConfigStore) CreateSuppressionTx(_ context.Context, _ pgx.Tx, _ *config.PurchaseSuppression) error {
-	return nil
-}
-func (m *MockConfigStore) CancelExecutionAtomic(ctx context.Context, tx pgx.Tx, executionID string, cancelledBy *string) (bool, string, error) {
-	args := m.Called(ctx, tx, executionID, cancelledBy)
-	return args.Bool(0), args.String(1), args.Error(2)
-}
-
-func (m *MockConfigStore) DeleteSuppressionsByExecution(_ context.Context, _ string) error {
-	return nil
-}
-func (m *MockConfigStore) DeleteSuppressionsByExecutionTx(ctx context.Context, tx pgx.Tx, executionID string) error {
-	return m.Called(ctx, tx, executionID).Error(0)
-}
-func (m *MockConfigStore) ListActiveSuppressions(_ context.Context) ([]config.PurchaseSuppression, error) {
-	return nil, nil
-}
-func (m *MockConfigStore) SavePurchaseExecutionTx(ctx context.Context, _ pgx.Tx, exec *config.PurchaseExecution) error {
-	// Forward to the plain variant so tests that only assert on
-	// SavePurchaseExecution still see the write.
-	return m.SavePurchaseExecution(ctx, exec)
-}
-func (m *MockConfigStore) GetPendingExecutionsTx(ctx context.Context, _ pgx.Tx) ([]config.PurchaseExecution, error) {
-	return m.GetPendingExecutions(ctx)
-}
-func (m *MockConfigStore) WithTx(_ context.Context, fn func(tx pgx.Tx) error) error {
-	return fn(nil)
-}
-
-func (m *MockConfigStore) ReplaceRecommendations(_ context.Context, _ time.Time, _ []config.RecommendationRecord) error {
-	return nil
-}
-func (m *MockConfigStore) UpsertRecommendations(_ context.Context, _ time.Time, _ []config.RecommendationRecord, _ []config.SuccessfulCollect) error {
-	return nil
-}
-func (m *MockConfigStore) ListStoredRecommendations(_ context.Context, _ config.RecommendationFilter) ([]config.RecommendationRecord, error) {
-	return nil, nil
-}
-func (m *MockConfigStore) GetRecommendationsFreshness(_ context.Context) (*config.RecommendationsFreshness, error) {
-	return &config.RecommendationsFreshness{}, nil
-}
-func (m *MockConfigStore) SetRecommendationsCollectionError(_ context.Context, _ string) error {
-	return nil
-}
-func (m *MockConfigStore) MarkCollectionStarted(_ context.Context) (bool, error) {
-	return true, nil
-}
-func (m *MockConfigStore) ClearCollectionStarted(_ context.Context) error {
-	return nil
-}
-func (m *MockConfigStore) GetRIUtilizationCache(_ context.Context, _ string, _ int) (*config.RIUtilizationCacheEntry, error) {
-	return nil, nil
-}
-func (m *MockConfigStore) UpsertRIUtilizationCache(_ context.Context, _ string, _ int, _ []byte, _ time.Time) error {
-	return nil
-}
-
-// Verify MockConfigStore implements config.StoreInterface
-var _ config.StoreInterface = (*MockConfigStore)(nil)
 
 // MockEmailSender is a mock implementation of email.SenderInterface
 type MockEmailSender struct {
